writerprediction/requirements.txt
streamlit
tensorflow
opencv-python
scikit-learn==1.2.2
pillow
scipy
scikit-image